fix(): 修复小数加购的bug

+++ app/javascripts/productDetails.js
@@ -313,7 +313,7 @@ window.App = {
     var totalPrice = Cookies.get('cart-price');
 
     ids = (ids == null)?(new Map()):(_objToStrMap(JSON.parse(ids)));
-    totalPrice = (totalPrice == null)?0:parseInt(totalPrice);
+    totalPrice = (totalPrice == null)?0:parseFloat(totalPrice);
     ids_size = (ids_size == null)?0:parseInt(ids_size);
 
     if (ids.get("article"+article_id) == null){
@@ -331,7 +331,7 @@ window.App = {
     Market.deployed().then(function (instance) {
       instance.articles(article_id).then(function (article) {
         var etherPrice = web3.fromWei(article[5], "ether");
-        totalPrice += parseInt(etherPrice);
+        totalPrice += parseFloat(etherPrice);
         document.getElementById("cartPrice").innerText = totalPrice+" ETH";
         Cookies.set('cart-price',totalPrice);
       })
@@ -344,7 +344,7 @@ window.App = {
     var totalPrice = Cookies.get('cart-price');
 
     ids = (ids == null)?(new Map()):(_objToStrMap(JSON.parse(ids)));
-    totalPrice = (totalPrice == null)?0:parseInt(totalPrice);
+    totalPrice = (totalPrice == null)?0:parseFloat(totalPrice);
     ids_size = (ids_size == null)?0:parseInt(ids_size);
 
     ids.forEach(function (value, key) {
